Add dynamic polyfills for older browsers
Fixes #2941
