+++ package.json
@@ -17,6 +17,7 @@
     "url": "https://github.com/tootsuite/mastodon.git"
   },
   "dependencies": {
+    "array-includes": "^3.0.3",
     "autoprefixer": "^6.7.7",
     "axios": "^0.15.3",
     "babel-cli": "^6.23.0",
@@ -47,6 +48,7 @@
     "emojione": "^2.2.7",
     "emojione-picker": "^2.2.1",
     "es6-promise": "^3.2.1",
+    "es6-symbol": "^3.1.1",
     "escape-html": "^1.0.3",
     "eventsource": "^0.2.1",
     "express": "^4.14.1",
@@ -57,10 +59,12 @@
     "http-link-header": "^0.8.0",
     "immutable": "^3.8.1",
     "intl": "^1.2.5",
+    "is-nan": "^1.2.1",
     "js-yaml": "^3.8.3",
     "lodash": "^4.17.4",
     "node-sass": "^4.5.2",
     "npmlog": "^4.0.2",
+    "object-assign": "^4.1.1",
     "path-complete-extname": "^0.1.0",
     "pg": "^6.1.2",
     "postcss-loader": "^1.3.3",
